requirements.txt
alembic==1.7.7
asyncpg==0.26.0
SQLAlchemy==2.0.1
uvicorn[standard]==0.18.0
fastapi==0.73.0
pydantic==1.9.0
psycopg2-binary==2.9.3

python-dotenv~=1.1.0
